+++ src/core/currency/wallet/metadata.ts
@@ -1,6 +1,6 @@
 import { asNumber, asObject, asOptional, asString, Cleaner } from 'cleaners'
 
-import { EdgeMetadata } from '../../../types/types'
+import { EdgeMetadata, EdgeMetadataChange } from '../../../types/types'
 import { CurrencyWalletInput } from './currency-wallet-pixie'
 
 export const asEdgeMetadata: Cleaner<EdgeMetadata> = raw => {
@@ -19,23 +19,34 @@ export const asEdgeMetadata: Cleaner<EdgeMetadata> = raw => {
 
 export function mergeMetadata(
   under: EdgeMetadata,
-  over: EdgeMetadata
+  over: EdgeMetadata | EdgeMetadataChange
 ): EdgeMetadata {
-  return {
-    exchangeAmount: {
-      ...under.exchangeAmount,
-      ...over.exchangeAmount
-    },
-    bizId: over.bizId ?? under.bizId,
-    category: over.category ?? under.category,
-    name: over.name ?? under.name,
-    notes: over.notes ?? under.notes
+  const out: EdgeMetadata = { exchangeAmount: {} }
+  const { exchangeAmount = {} } = out
+
+  // Merge the fiat amounts:
+  const underAmounts = under.exchangeAmount ?? {}
+  const overAmounts = over.exchangeAmount ?? {}
+  for (const fiat of Object.keys(underAmounts)) {
+    if (overAmounts[fiat] !== null) exchangeAmount[fiat] = underAmounts[fiat]
   }
+  for (const fiat of Object.keys(overAmounts)) {
+    const amount = overAmounts[fiat]
+    if (amount != null) exchangeAmount[fiat] = amount
+  }
+
+  // Merge simple fields:
+  if (over.bizId !== null) out.bizId = over.bizId ?? under.bizId
+  if (over.category !== null) out.category = over.category ?? under.category
+  if (over.name !== null) out.name = over.name ?? under.name
+  if (over.notes !== null) out.notes = over.notes ?? under.notes
+
+  return out
 }
 
 export function upgradeMetadata(
   input: CurrencyWalletInput,
-  metadata: EdgeMetadata
+  metadata: EdgeMetadata | EdgeMetadataChange
 ): void {
   const { fiat = 'iso:USD' } = input.props.walletState
   if (metadata.amountFiat != null) {